refactor: :recycle: useModal 없앰, 모달 코드 리팩토링, handleCloseModal 외부에서 받아오게 수정
