+++ frontend/src/hooks/index.ts
@@ -1,4 +1,3 @@
 export { useHandleSocket } from './useHandleSocket';
-export { useModal } from './useModal';
 export { useGetCurrentUser } from './useGetCurrentUser';
 export { useLogout } from './useLogout';
