--- frontend/src/pages/ChatPage/ChatInfoModal/ChatInfoModal.styles.ts
@@ -1,5 +1,0 @@
-import { css } from '@emotion/css';
-
-export const closeButtonStyle = css`
-  border: 1px solid black; // TODO: 색상 상수화
-`;
