+++ frontend/src/pages/ChatPage/ChatInfoModalHeader.styles.ts
